refactor(device): extract TrainDevice HW/sim layering
Share train hub Properties via TrainDevice, factor BLE into
BleDevice for later switch hubs, and keep Train device-agnostic.

Refs #154
