+++ Controller/TrainsController.pyproject
@@ -8,6 +8,7 @@
         "tests/test_network_manager.py",
         "tests/test_marker_sanity.py",
         "tests/test_simulator_pause.py",
+        "tests/test_train_device.py",
         "src/main.py",
         "src/projects/rails.json",
         "src/resources/rails_rc.py",
@@ -38,8 +39,10 @@
         "src/python/models/trains.py",
         "src/python/items/rotator.py",
         "src/python/items/connector.py",
-        "src/python/items/device.py",
-        "src/python/items/fake_device.py",
+        "src/python/items/ble_device.py",
+        "src/python/items/train_device.py",
+        "src/python/items/train_device_hw.py",
+        "src/python/items/train_device_sim.py",
         "src/python/items/marker.py",
         "src/python/items/marker_type.py",
         "src/python/items/rail.py",
